feat: move properties from code behind
Move style properties from code behind for controls to special 'properties' subclasses

Asana: https://app.asana.com/1/1200746044236722/project/1203851531040615/task/1209709314147235?focus=true
