+++ src/Asv.Avalonia.GeoMap/Controls/Canvas/AnnotationLayer.cs
@@ -12,7 +12,7 @@ using R3;
 
 namespace Asv.Avalonia.GeoMap;
 
-public class AnnotationLayer : Canvas
+public partial class AnnotationLayer : Canvas
 {
     static AnnotationLayer()
     {
@@ -45,59 +45,6 @@ public class AnnotationLayer : Canvas
         _annotations.Clear();
         base.OnUnloaded(e);
     }
-
-    #region ItemTemplate
-
-    public static readonly StyledProperty<IDataTemplate?> ItemTemplateProperty =
-        AvaloniaProperty.Register<AnnotationLayer, IDataTemplate?>(nameof(ItemTemplate));
-
-    public IDataTemplate? ItemTemplate
-    {
-        get => GetValue(ItemTemplateProperty);
-        set => SetValue(ItemTemplateProperty, value);
-    }
-
-    #endregion
-
-    #region Connector line
-
-    public static readonly StyledProperty<double> StrokeThicknessProperty =
-        AvaloniaProperty.Register<AnnotationLayer, double>(
-            nameof(StrokeThickness),
-            defaultValue: 1
-        );
-
-    public double StrokeThickness
-    {
-        get => GetValue(StrokeThicknessProperty);
-        set => SetValue(StrokeThicknessProperty, value);
-    }
-
-    public static readonly StyledProperty<IBrush?> StrokeProperty = AvaloniaProperty.Register<
-        AnnotationLayer,
-        IBrush?
-    >(nameof(Stroke), defaultValue: Brushes.Violet);
-
-    public IBrush? Stroke
-    {
-        get => GetValue(StrokeProperty);
-        set => SetValue(StrokeProperty, value);
-    }
-
-    #endregion
-
-    #region Source
-
-    public static readonly StyledProperty<MapItemsControl?> SourceProperty =
-        AvaloniaProperty.Register<AnnotationLayer, MapItemsControl?>(nameof(Source));
-
-    public MapItemsControl? Source
-    {
-        get => GetValue(SourceProperty);
-        set => SetValue(SourceProperty, value);
-    }
-
-    #endregion
 
     private void MapControlSourceUpdated(AvaloniaPropertyChangedEventArgs<MapItemsControl?> e)
     {
