+++ src/Asv.Avalonia/Controls/Dock/DockControl.axaml.cs
@@ -4,11 +4,9 @@ using Asv.Common;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
-using Avalonia.Controls.Templates;
 using Avalonia.Input;
 using Avalonia.Interactivity;
 using Avalonia.Media;
-using Avalonia.Metadata;
 using Avalonia.VisualTree;
 using R3;
 
@@ -21,7 +19,7 @@ public class ShellItem
     public int Column { get; set; }
 }
 
-public class DockControl : SelectingItemsControl
+public partial class DockControl : SelectingItemsControl
 {
     private const int ColumnIncrement = 2;
     private readonly List<Border> _targetBorders = [];
@@ -600,52 +598,6 @@ public class DockControl : SelectingItemsControl
         {
             pair.curr.Column -= ColumnIncrement;
         }
-    }
-
-    #endregion
-
-    #region Properties
-
-    public static StyledProperty<ReactiveCommand> UnSplitAllCommandProperty =
-        AvaloniaProperty.Register<DockControl, ReactiveCommand>(nameof(UnSplitAllCommand));
-
-    public ReactiveCommand UnSplitAllCommand
-    {
-        get => GetValue(UnSplitAllCommandProperty);
-        set => SetValue(UnSplitAllCommandProperty, value);
-    }
-
-    public static readonly StyledProperty<IDataTemplate?> TabControlStripItemTemplateProperty =
-        AvaloniaProperty.Register<DockControl, IDataTemplate?>(nameof(TabControlStripItemTemplate));
-
-    [InheritDataTypeFromItems("ItemsSource")]
-    public IDataTemplate? TabControlStripItemTemplate
-    {
-        get => GetValue(TabControlStripItemTemplateProperty);
-        set => SetValue(TabControlStripItemTemplateProperty, value);
-    }
-
-    public static readonly StyledProperty<int> MaxSplitAmountProperty = AvaloniaProperty.Register<
-        DockControl,
-        int
-    >(nameof(MaxSplitAmount), 4);
-
-    public int MaxSplitAmount
-    {
-        get => GetValue(MaxSplitAmountProperty);
-        set => SetValue(MaxSplitAmountProperty, value);
-    }
-
-    public static readonly StyledProperty<IBrush> BorderHighLightColorProperty =
-        AvaloniaProperty.Register<DockControl, IBrush>(
-            nameof(BorderHighLightColor),
-            Brushes.LightBlue
-        );
-
-    public IBrush BorderHighLightColor
-    {
-        get => GetValue(BorderHighLightColorProperty);
-        set => SetValue(BorderHighLightColorProperty, value);
     }
 
     #endregion
