+++ src/Asv.Avalonia/Controls/FrequencyScreen/FrequencyControl.axaml.cs
@@ -1,18 +1,5 @@
-﻿using Avalonia;
-using Avalonia.Controls.Primitives;
+﻿using Avalonia.Controls.Primitives;
 
 namespace Asv.Avalonia;
 
-public class FrequencyControl : TemplatedControl
-{
-    public static readonly StyledProperty<string?> ValueProperty = AvaloniaProperty.Register<
-        FrequencyControl,
-        string?
-    >(nameof(Value));
-
-    public string? Value
-    {
-        get => GetValue(ValueProperty);
-        set => SetValue(ValueProperty, value);
-    }
-}
+public partial class FrequencyControl : TemplatedControl { }
