+++ src/Asv.Avalonia/Controls/Indicators/ConnectionQuality.axaml.cs
@@ -11,68 +11,8 @@ namespace Asv.Avalonia;
     PseudoClassesHelper.Normal,
     PseudoClassesHelper.Unknown
 )]
-public class ConnectionQuality : IndicatorBase
+public partial class ConnectionQuality : IndicatorBase
 {
-    #region Styled Props
-
-    public static readonly StyledProperty<double> CriticalValueProperty = AvaloniaProperty.Register<
-        ConnectionQuality,
-        double
-    >(nameof(CriticalValue), 0.2);
-
-    public double CriticalValue
-    {
-        get => GetValue(CriticalValueProperty);
-        set => SetValue(CriticalValueProperty, value);
-    }
-
-    public static readonly StyledProperty<double> WarningValueProperty = AvaloniaProperty.Register<
-        ConnectionQuality,
-        double
-    >(nameof(WarningValue), 0.5);
-
-    public double WarningValue
-    {
-        get => GetValue(WarningValueProperty);
-        set => SetValue(WarningValueProperty, value);
-    }
-
-    public static readonly StyledProperty<double> MaxValueProperty = AvaloniaProperty.Register<
-        ConnectionQuality,
-        double
-    >(nameof(MaxValue), 1);
-
-    public double MaxValue
-    {
-        get => GetValue(MaxValueProperty);
-        set => SetValue(MaxValueProperty, value);
-    }
-
-    public static readonly StyledProperty<double?> ValueProperty = AvaloniaProperty.Register<
-        ConnectionQuality,
-        double?
-    >(nameof(Value), default(double?));
-
-    public double? Value
-    {
-        get => GetValue(ValueProperty);
-        set => SetValue(ValueProperty, value);
-    }
-
-    public static readonly StyledProperty<MaterialIconKind> IconKindProperty =
-        AvaloniaProperty.Register<ConnectionQuality, MaterialIconKind>(
-            nameof(IconKind),
-            MaterialIconKind.WifiStrengthAlertOutline
-        );
-
-    public MaterialIconKind IconKind
-    {
-        get => GetValue(IconKindProperty);
-        set => SetValue(IconKindProperty, value);
-    }
-
-    #endregion
-
     public ConnectionQuality() { }
 
     protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change)
