+++ src/Asv.Avalonia/Controls/Indicators/GpsStatusIndicator.axaml.cs
@@ -82,53 +82,8 @@ public enum DopStatusEnum
     PseudoClassesHelper.Normal,
     PseudoClassesHelper.Unknown
 )]
-public class GpsStatusIndicator : IndicatorBase
+public partial class GpsStatusIndicator : IndicatorBase
 {
-    #region Styled Props
-
-    public static readonly StyledProperty<GpsFixType> FixTypeProperty = AvaloniaProperty.Register<
-        GpsStatusIndicator,
-        GpsFixType
-    >(nameof(FixType), GpsFixType.GpsFixTypeNoGps, coerce: UpdateValue);
-
-    public GpsFixType FixType
-    {
-        get => GetValue(FixTypeProperty);
-        set => SetValue(FixTypeProperty, value);
-    }
-
-    public static readonly StyledProperty<string> ToolTipTextProperty = AvaloniaProperty.Register<
-        GpsStatusIndicator,
-        string
-    >(nameof(ToolTipText), string.Empty);
-
-    public string ToolTipText
-    {
-        get => GetValue(ToolTipTextProperty);
-        set => SetValue(ToolTipTextProperty, value);
-    }
-
-    public static readonly StyledProperty<DopStatusEnum> DopStatusProperty =
-        AvaloniaProperty.Register<GpsStatusIndicator, DopStatusEnum>(
-            nameof(DopStatus),
-            DopStatusEnum.Unknown,
-            coerce: UpdateValue
-        );
-
-    private static DopStatusEnum UpdateValue(AvaloniaObject arg1, DopStatusEnum arg2)
-    {
-        // TODO: AVALONIA11 => implement this
-        return DopStatusEnum.Excellent;
-    }
-
-    public DopStatusEnum DopStatus
-    {
-        get => GetValue(DopStatusProperty);
-        set => SetValue(DopStatusProperty, value);
-    }
-
-    #endregion
-
     public static void SetPseudoClass(GpsStatusIndicator indicator)
     {
         var dopStatus = indicator.DopStatus;
