+++ src/Asv.Avalonia/Controls/Workspace/WorkspacePanel.properties.cs
